Extract GitHub OAuth auth routes
